[ivp_intern] Harden memory initialization with guarded scratchpad setup for IVP_Environment
Replace unconditional scratchpad memory assignment with defensive
initialization that validates pointer and size constraints. Memory
regions are now only utilized when they meet IVU_MEM_ALIGN minimum
requirements, falling back to heap allocation otherwise.

Additional improvements:
- Promote scratchpad_size to size_t for 64-bit address space compatibility
- Consolidate integrated_energy_damp calculation into set_delta_PSI_time
- Remove orphaned debug_text free in draw vector cleanup

+++ ivp_physics/ivp_environment.hxx
@@ -441,7 +441,7 @@ public:
     int n_cache_object; // == 2**x number of caches for moving objects, should be larger than typical number of moving objects
 
     char *scratchpad_addr; // scratch pad: a fast area which is temporarily used and is already cached in
-    int scratchpad_size;
+    size_t scratchpad_size;
 
     /********************************************************************************
      *	Name:	     	Several custom managers
